requirements.txt
PyQt5==5.15.10
Pillow>=10.2.0
numpy>=1.26.0
